Even more porting work

+++ src/main/java/com/mjr/extraplanets/planets/Kepler22b/worldgen/ChunkProviderKepler22b.java
@@ -7,17 +7,11 @@ import micdoodle8.mods.galacticraft.core.perlin.generator.Gradient;
 import net.minecraft.block.BlockFalling;
 import net.minecraft.entity.EnumCreatureType;
 import net.minecraft.init.Blocks;
-import net.minecraft.util.BlockPos;
 import net.minecraft.util.IProgressUpdate;
-import net.minecraft.util.MathHelper;
-import net.minecraft.world.ChunkCoordIntPair;
-import net.minecraft.world.SpawnerAnimals;
 import net.minecraft.world.World;
-import net.minecraft.world.biome.BiomeGenBase;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.ChunkPrimer;
 import net.minecraft.world.chunk.IChunkProvider;
-import net.minecraft.world.gen.ChunkProviderGenerate;
 import net.minecraft.world.gen.NoiseGeneratorOctaves;
 import net.minecraft.world.gen.NoiseGeneratorPerlin;
 
@@ -72,7 +66,7 @@ public class ChunkProviderKepler22b extends ChunkProviderGenerate {
 
 	@Override
 	public Chunk provideChunk(int chunkX, int chunkZ) {
-		this.rand.setSeed((long) chunkX * 341873128712L + (long) chunkZ * 132897987541L);
+		this.rand.setSeed(chunkX * 341873128712L + chunkZ * 132897987541L);
 		ChunkPrimer chunkprimer = new ChunkPrimer();
 		this.setBlocksInChunk(chunkX, chunkZ, chunkprimer);
 		this.biomesForGeneration = this.worldObj.getWorldChunkManager().loadBlockGeneratorData(this.biomesForGeneration, chunkX * 16, chunkZ * 16, 16, 16);
@@ -153,8 +147,8 @@ public class ChunkProviderKepler22b extends ChunkProviderGenerate {
 		float f = 684.412F;
 		float f1 = 684.412F;
 		this.field_147427_d = this.field_147429_l.generateNoiseOctaves(this.field_147427_d, chunkX, chunkY, chunkZ, 5, 33, 5, f / 80.0F, f1 / 160.0F, f / 80.0F);
-		this.field_147428_e = this.field_147431_j.generateNoiseOctaves(this.field_147428_e, chunkX, chunkY, chunkZ, 5, 33, 5, (double) f, (double) f1, (double) f);
-		this.field_147425_f = this.field_147432_k.generateNoiseOctaves(this.field_147425_f, chunkX, chunkY, chunkZ, 5, 33, 5, (double) f, (double) f1, (double) f);
+		this.field_147428_e = this.field_147431_j.generateNoiseOctaves(this.field_147428_e, chunkX, chunkY, chunkZ, 5, 33, 5, f, f1, f);
+		this.field_147425_f = this.field_147432_k.generateNoiseOctaves(this.field_147425_f, chunkX, chunkY, chunkZ, 5, 33, 5, f, f1, f);
 		chunkZ = 0;
 		chunkX = 0;
 		int i = 0;
@@ -215,14 +209,14 @@ public class ChunkProviderKepler22b extends ChunkProviderGenerate {
 				}
 
 				++j;
-				double d8 = (double) f3;
-				double d9 = (double) f2;
+				double d8 = f3;
+				double d9 = f2;
 				d8 = d8 + d7 * 0.2D;
 				d8 = d8 * 8.5F / 8.0D;
 				double d0 = 8.5F + d8 * 4.0D;
 
 				for (int l1 = 0; l1 < 33; ++l1) {
-					double d1 = ((double) l1 - d0) * 12.0F * 128.0D / 256.0D / d9;
+					double d1 = (l1 - d0) * 12.0F * 128.0D / 256.0D / d9;
 
 					if (d1 < 0.0D) {
 						d1 *= 4.0D;
@@ -234,7 +228,7 @@ public class ChunkProviderKepler22b extends ChunkProviderGenerate {
 					double d5 = MathHelper.denormalizeClamp(d2, d3, d4) - d1;
 
 					if (l1 > 29) {
-						double d6 = (double) ((float) (l1 - 29) / 3.0F);
+						double d6 = (l1 - 29) / 3.0F;
 						d5 = d5 * (1.0D - d6) + -10.0D * d6;
 					}
